Create compute_commission_report.py
Constants Setup
The program starts by defining the commission values for each apartment type and a special QUIT value (9999) that the user can enter to stop the program.

Get First Salesperson ID
The user is asked to enter a salesperson ID. If they enter 9999, the program ends immediately. Otherwise, the program continues.

Main Loop
While the ID entered is not 9999, the program does the following:

Prompts the user for the salesperson’s name.

Prompts the user for the number of bedrooms rented.

Determines the commission based on the number of bedrooms using an if-elif-else structure.

Displays a report showing the salesperson’s ID, name, and the commission earned.

Asks again for the next salesperson ID (or 9999 to quit).

End of Report
Once the user enters 9999 as the salesperson ID, the program prints "End of report" and stops.